Group the module lifecycle behind a Modules command
The CLI now presents modules as one noun (`module install/update/remove/
list`) while the same four operations were spread over Import, Remove and
Show. Give them a single owner.

`Modules` takes install/update/updateAll from `Import` (which is left with
CSV card import), `removeModule` from `Remove`, and the two module reads
from `Show`. `Remove` delegates its `module` branch; `Show.showProject` and
`Modules.list` share a new `moduleInfos` in the inventory layer, so `Show`
keeps its single-argument constructor.

The `Cmd` dispatcher verbs are unchanged — `Cmd.import`/`remove`/`show`/
`update-modules` still route the same calls, just to the new owner. Module
tests move out of command-import.test.ts into command-modules.test.ts with
their own temp dir.

No behaviour change: the write-lock descriptions and every error string are
carried over verbatim.
