+++ tools/data-handler/src/command-manager.ts
@@ -18,6 +18,7 @@ import { Edit } from './commands/edit.js';
 import { Export } from './commands/export.js';
 import { Fetch } from './commands/fetch.js';
 import { Import } from './commands/import.js';
+import { Modules } from './commands/modules.js';
 import { Move } from './commands/move.js';
 import { Publish } from './commands/publish.js';
 import { Remove } from './commands/remove.js';
@@ -54,6 +55,7 @@ export class CommandManager {
   public exportCmd: Export;
   public fetchCmd: Fetch;
   public importCmd: Import;
+  public modulesCmd: Modules;
   public moveCmd: Move;
   public publishCmd: Publish;
   public removeCmd: Remove;
@@ -81,14 +83,15 @@ export class CommandManager {
     this.checkUpdatesCmd = new CheckUpdates(this.project);
     this.cleanCmd = new Clean(this.project);
     this.fetchCmd = new Fetch(this.project);
+    this.modulesCmd = new Modules(this.project, this.fetchCmd);
     this.showCmd = new Show(this.project);
     this.createCmd = new Create(this.project);
     this.editCmd = new Edit(this.project);
     this.exportCmd = new Export(this.project, this.showCmd);
-    this.importCmd = new Import(this.project, this.createCmd, this.fetchCmd);
+    this.importCmd = new Import(this.project, this.createCmd);
     this.moveCmd = new Move(this.project);
     this.publishCmd = new Publish(this.project);
-    this.removeCmd = new Remove(this.project, this.fetchCmd);
+    this.removeCmd = new Remove(this.project, this.fetchCmd, this.modulesCmd);
     this.renameCmd = new Rename(this.project);
     this.transitionCmd = new Transition(this.project);
     this.updateCmd = new Update(this.project);
