+++ tools/data-handler/test/command-modules.test.ts
@@ -1,0 +1,1152 @@
+import {
+  expect,
+  it,
+  describe,
+  beforeAll,
+  beforeEach,
+  afterEach,
+  vi,
+} from 'vitest';
+
+import {
+  existsSync,
+  mkdirSync,
+  readFileSync,
+  rmSync,
+  writeFileSync,
+} from 'node:fs';
+import { join, resolve as pathResolve } from 'node:path';
+
+import { CommandManager } from '../src/command-manager.js';
+import { copyDir } from '../src/utils/file-utils.js';
+import { Cmd, Commands } from '../src/command-handler.js';
+import { GitManager } from '../src/utils/git-manager.js';
+import {
+  makeFakeModuleFixture,
+  rewriteFakeModuleFixture,
+} from './helpers/module-fixtures.js';
+import { mockEnsureModuleListUpToDate } from './helpers/test-utils.js';
+import { readJsonFileSync } from '../src/utils/json.js';
+import { SCHEMA_VERSION } from '@cyberismo/migrations';
+import { toVersionRange } from '../src/modules/types.js';
+import { ModuleValidationFailedError } from '../src/mutations/replay/replay.js';
+import { logLine, writeSeals } from './helpers/replay-fixtures.js';
+
+// Create test artifacts in a temp folder.
+const baseDir = import.meta.dirname;
+const testDir = join(baseDir, 'tmp-command-handler-modules-tests');
+
+const decisionRecordsPath = join(testDir, 'valid/decision-records');
+const minimalPath = join(testDir, 'valid/minimal');
+
+const commandHandler: Commands = new Commands();
+const optionsMini = { projectPath: minimalPath };
+const options = { projectPath: decisionRecordsPath };
+
+describe('import module', () => {
+  beforeEach(async () => {
+    mkdirSync(testDir, { recursive: true });
+    await copyDir('test/test-data', testDir);
+  });
+
+  afterEach(() => {
+    rmSync(testDir, { recursive: true, force: true });
+    vi.restoreAllMocks();
+  });
+
+  describe('import module command', () => {
+    it('import module and use it (success)', async () => {
+      let result = await commandHandler.command(
+        Cmd.import,
+        ['module', decisionRecordsPath],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(200);
+
+      // Verify that module content can be used to create data.
+      result = await commandHandler.command(
+        Cmd.create,
+        ['cardType', 'newCardType', 'decision/workflows/decision'],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(200);
+
+      // Ensure that module can be updated.
+      result = await commandHandler.command(Cmd.updateModules, [], optionsMini);
+      expect(result.statusCode).toBe(200);
+
+      // Remove the module so that it won't affect other tests
+      await commandHandler.command(
+        Cmd.remove,
+        ['module', 'decision'],
+        optionsMini,
+      );
+    });
+    it('create empty project and import two modules', async () => {
+      const prefix = 'proj';
+      const name = 'test-project';
+      const projectDir = join(testDir, name);
+      const testOptions = { projectPath: projectDir };
+      const data = await commandHandler.command(
+        Cmd.create,
+        ['project', name, prefix],
+        testOptions,
+      );
+      expect(data.statusCode).toBe(200);
+      let result = await commandHandler.command(
+        Cmd.import,
+        ['module', decisionRecordsPath],
+        testOptions,
+      );
+      expect(result.statusCode).toBe(200);
+      result = await commandHandler.command(
+        Cmd.import,
+        ['module', minimalPath],
+        testOptions,
+      );
+      expect(result.statusCode).toBe(200);
+      result = await commandHandler.command(Cmd.updateModules, [], testOptions);
+      expect(result.statusCode).toBe(200);
+      result = await commandHandler.command(Cmd.show, ['modules'], testOptions);
+      expect(result.statusCode).toBe(200);
+      expect(result.payload).toBeDefined();
+      const modules = result.payload as Array<{ name: string }>;
+      expect(modules.length).toBe(2);
+      expect(modules.map((m) => m.name)).toContain('mini');
+      expect(modules.map((m) => m.name)).toContain('decision');
+    }, 10000);
+    it('try to import module - no source', async () => {
+      const stubProjectPath = vi
+        .spyOn(commandHandler, 'setProjectPath')
+        .mockResolvedValue('path');
+      const result = await commandHandler.command(
+        Cmd.import,
+        ['module', ''],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(400);
+      stubProjectPath.mockRestore();
+    });
+    it('try to import module - no destination', async () => {
+      const stubProjectPath = vi
+        .spyOn(commandHandler, 'setProjectPath')
+        .mockResolvedValue('path');
+      const invalidOptions = { projectPath: '' };
+      await expect(
+        commandHandler.command(
+          Cmd.import,
+          ['module', decisionRecordsPath],
+          invalidOptions,
+        ),
+      ).resolves.toEqual({
+        statusCode: 400,
+        message: "Input validation error: cannot find project ''",
+      });
+      stubProjectPath.mockRestore();
+    });
+    it('re-importing the same module is upsert', async () => {
+      // Re-importing a module that is already declared with the same
+      // source location must succeed and update the declared range
+      // rather than error.
+      const result1 = await commandHandler.command(
+        Cmd.import,
+        ['module', decisionRecordsPath],
+        optionsMini,
+      );
+      expect(result1.statusCode).toBe(200);
+      const result2 = await commandHandler.command(
+        Cmd.import,
+        ['module', decisionRecordsPath],
+        optionsMini,
+      );
+      expect(result2.statusCode).toBe(200);
+    });
+    it('re-importing with a mismatched source location is rejected', async () => {
+      // Once a module has been imported from a given source, a re-import
+      // of the same module name from a different source must be rejected.
+      // The resolver enforces this via `assertSourceAgreement`.
+      const firstImport = await commandHandler.command(
+        Cmd.import,
+        ['module', decisionRecordsPath],
+        optionsMini,
+      );
+      expect(firstImport.statusCode).toBe(200);
+
+      // Capture the persisted declaration so we can assert it is not
+      // mutated by the failed re-import.
+      const configPath = join(
+        minimalPath,
+        '.cards',
+        'local',
+        'cardsConfig.json',
+      );
+      const configBefore = readFileSync(configPath, 'utf-8');
+
+      // Make a sibling copy of the same module fixture at a different
+      // path. Same `cardKeyPrefix` ("decision"), different `file:` URL.
+      const altModulePath = join(testDir, 'valid/decision-records-alt');
+      await copyDir(decisionRecordsPath, altModulePath);
+
+      const result = await commandHandler.command(
+        Cmd.import,
+        ['module', altModulePath],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(400);
+      expect(result.message).toMatch(
+        /Conflicting source for module 'decision'/,
+      );
+
+      // Config must not have been mutated — the resolver's source-
+      // agreement check fires before the applier persists anything.
+      const configAfter = readFileSync(configPath, 'utf-8');
+      expect(configAfter).toBe(configBefore);
+    });
+    it('try to import module - that has the same prefix', async () => {
+      const result = await commandHandler.command(
+        Cmd.import,
+        ['module', minimalPath],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(400);
+    });
+    it('try to import module - no cardKeyPrefix in cardsConfig.json', async () => {
+      const noPrefix = join(testDir, 'fake-no-prefix');
+      makeFakeModuleFixture(noPrefix, { cardKeyPrefix: 'placeholder' });
+      writeFileSync(
+        join(noPrefix, '.cards', 'local', 'cardsConfig.json'),
+        JSON.stringify(
+          { cardKeyPrefix: '', name: 'noprefix', modules: [], hubs: [] },
+          null,
+          2,
+        ),
+      );
+      const result = await commandHandler.command(
+        Cmd.import,
+        ['module', noPrefix],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(400);
+      expect(result.message).toContain('cardKeyPrefix');
+    });
+  });
+
+  describe('modifying imported module content is forbidden', () => {
+    beforeAll(async () => {
+      await commandHandler.command(
+        Cmd.import,
+        ['module', minimalPath],
+        options,
+      );
+      await commandHandler.command(
+        Cmd.import,
+        ['module', decisionRecordsPath],
+        optionsMini,
+      );
+    });
+    it('try to add card to module template', async () => {
+      const templateName = 'mini/templates/test-template';
+      const cardType = 'decision/cardTypes/decision';
+      const cardKey = '';
+      const result = await commandHandler.command(
+        Cmd.add,
+        ['card', templateName, cardType, cardKey],
+        options,
+      );
+      expect(result.statusCode).toBe(400);
+    });
+    it('try to add child card to a module card', async () => {
+      const templateName = 'decision/templates/decision';
+      const cardType = 'decision/cardTypes/decision';
+      const cardKey = 'decision_2';
+      // try to add new card to decision_2 when 'decision-records' has been imported to 'minimal'
+      const result = await commandHandler.command(
+        Cmd.add,
+        ['card', templateName, cardType, cardKey],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(400);
+    });
+    it('try to create attachment to a module card', async () => {
+      const attachmentPath = join(testDir, 'attachments/the-needle.heic');
+      const cardKey = 'decision_2';
+      const result = await commandHandler.command(
+        Cmd.create,
+        ['attachment', cardKey, attachmentPath],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(400);
+    });
+
+    it('try to move a module card to another template', async () => {
+      const moduleCardKey = 'decision_2';
+      const templateCardKey = 'decision_1';
+      const result = await commandHandler.command(
+        Cmd.move,
+        [templateCardKey, moduleCardKey, 'root'],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(400);
+    });
+    it('try to remove card from a module template', async () => {
+      const cardKey = 'decision_2';
+      const result = await commandHandler.command(
+        Cmd.remove,
+        ['card', cardKey],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(400);
+    });
+    it('try to remove cardType from a module', async () => {
+      const cardType = 'decision/cardTypes/decision';
+      const result = await commandHandler.command(
+        Cmd.remove,
+        ['cardType', cardType],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(400);
+    });
+    it('try to remove fieldType from a module', async () => {
+      const fieldType = 'decision/fieldTypes/finished';
+      const result = await commandHandler.command(
+        Cmd.remove,
+        ['fieldType', fieldType],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(400);
+    });
+    it('try to remove report from a module', async () => {
+      const report = 'decision/reports/testReport';
+      const result = await commandHandler.command(
+        Cmd.remove,
+        ['report', report],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(400);
+    });
+    it('try to remove template from a module', async () => {
+      const template = 'decision/templates/decision';
+      const result = await commandHandler.command(
+        Cmd.remove,
+        ['template', template],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(400);
+    });
+    it('try to remove workflow from a module', async () => {
+      const workflow = 'decision/workflows/decision';
+      const result = await commandHandler.command(
+        Cmd.remove,
+        ['workflow', workflow],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(400);
+    });
+    it('try to remove attachment from a module card', async () => {
+      const cardKey = 'decision_1';
+      const attachment = 'the-needle.heic';
+      const result = await commandHandler.command(
+        Cmd.remove,
+        ['attachment', cardKey, attachment],
+        optionsMini,
+      );
+      expect(result.statusCode).toBe(400);
+    });
+  });
+});
+
+describe('update-modules version arg', () => {
+  beforeEach(async () => {
+    mkdirSync(testDir, { recursive: true });
+    await copyDir('test/test-data', testDir);
+  });
+
+  afterEach(() => {
+    rmSync(testDir, { recursive: true, force: true });
+    vi.restoreAllMocks();
+  });
+
+  it('version without module name returns error', async () => {
+    const result = await commandHandler.command(
+      Cmd.updateModules,
+      ['', '1.0.0'],
+      optionsMini,
+    );
+    expect(result.statusCode).toBe(400);
+    expect(result.message).toContain(
+      'A target version can only be specified together with a module name',
+    );
+  });
+
+  it('version with unknown module name returns error', async () => {
+    mockEnsureModuleListUpToDate();
+    const result = await commandHandler.command(
+      Cmd.updateModules,
+      ['nonexistent-module', '1.0.0'],
+      optionsMini,
+    );
+    expect(result.statusCode).toBe(400);
+    expect(result.message).toContain(
+      "Module 'nonexistent-module' is not part of the project",
+    );
+  });
+
+  it('version not in available list returns error', async () => {
+    // Import a local module so it appears in cardsConfig.json, then flip
+    // the persisted location to a git URL so the version check is routed
+    // through the source layer (which only queries tags for git remotes).
+    await commandHandler.command(
+      Cmd.import,
+      ['module', decisionRecordsPath],
+      optionsMini,
+    );
+
+    const configPath = join(minimalPath, '.cards', 'local', 'cardsConfig.json');
+    const config = JSON.parse(readFileSync(configPath, 'utf-8'));
+    const decisionModule = config.modules?.find(
+      (m: { name: string }) => m.name === 'decision',
+    );
+    if (decisionModule) {
+      decisionModule.location = 'https://example.com/decision.git';
+      writeFileSync(configPath, JSON.stringify(config, null, 4));
+    }
+
+    // Force the CommandManager singleton to reload from disk by routing
+    // through a different project path first — otherwise the in-memory
+    // Project still holds the module location from before the rewrite above.
+    await commandHandler.command(Cmd.show, ['project'], {
+      projectPath: decisionRecordsPath,
+    });
+
+    mockEnsureModuleListUpToDate();
+    // `Modules.update` consults the remote via the source layer, which in
+    // turn delegates to `GitManager.listRemoteVersionTags`.
+    vi.spyOn(GitManager, 'listRemoteVersionTags').mockResolvedValue([
+      '3.0.0',
+      '2.0.0',
+    ]);
+
+    const result = await commandHandler.command(
+      Cmd.updateModules,
+      ['decision', '1.0.0'],
+      optionsMini,
+    );
+    expect(result.statusCode).toBe(400);
+    expect(result.message).toContain("Version '1.0.0' is not available");
+    expect(result.message).toContain('2.0.0');
+    expect(result.message).toContain('3.0.0');
+  });
+});
+
+describe('module update — spec behaviours', () => {
+  const moduleTestDir = join(baseDir, 'tmp-command-import-module-update-tests');
+
+  beforeEach(async () => {
+    mkdirSync(moduleTestDir, { recursive: true });
+    await copyDir('test/test-data', moduleTestDir);
+  });
+
+  afterEach(() => {
+    rmSync(moduleTestDir, { recursive: true, force: true });
+    vi.restoreAllMocks();
+  });
+
+  it('updateAll cleans up installations orphaned by a dropped transitive dep', async () => {
+    // Build two fake modules: `host` (top-level) declares `dep`
+    // (transitive). Import host → dep is installed transitively. Edit
+    // host's fixture to drop dep, then run updateAll. After the
+    // orphan cascade, .cards/modules/dep/ must be gone.
+    const depRoot = join(moduleTestDir, 'fake-dep');
+    makeFakeModuleFixture(depRoot, { cardKeyPrefix: 'fkdep' });
+    const hostRoot = join(moduleTestDir, 'fake-host');
+    makeFakeModuleFixture(hostRoot, {
+      cardKeyPrefix: 'fkhost',
+      modules: [{ name: 'fkdep', location: `file:${pathResolve(depRoot)}` }],
+    });
+
+    const projectDir = join(moduleTestDir, 'proj-orphan');
+    const commandHandler = new Commands();
+    const create = await commandHandler.command(
+      Cmd.create,
+      ['project', 'orphan-proj', 'orph'],
+      { projectPath: projectDir },
+    );
+    expect(create.statusCode).toBe(200);
+
+    const commands = new CommandManager(projectDir, {});
+    await commands.initialize();
+
+    await commands.modulesCmd.install(hostRoot);
+
+    const installedHost = join(projectDir, '.cards', 'modules', 'fkhost');
+    const installedDep = join(projectDir, '.cards', 'modules', 'fkdep');
+    expect(existsSync(installedHost)).toBe(true);
+    expect(existsSync(installedDep)).toBe(true);
+
+    // Upstream drops its transitive dep.
+    rewriteFakeModuleFixture(hostRoot, {
+      cardKeyPrefix: 'fkhost',
+      modules: [],
+    });
+
+    await commands.modulesCmd.updateAll();
+
+    // host is still around; dep must have been removed by the orphan
+    // cascade that runs at the end of updateAll.
+    expect(existsSync(installedHost)).toBe(true);
+    expect(existsSync(installedDep)).toBe(false);
+  });
+
+  it('install cleans up installations orphaned by a dropped transitive dep', async () => {
+    // Running `Modules.install` on a host whose upstream dropped a transitive
+    // dep must remove the stale installation from disk. Transitives are
+    // not top-level declarations so they do not appear in
+    // `configuration.modules`; the proof is purely disk-based.
+    const depRoot = join(moduleTestDir, 'fake-dep-reimport');
+    makeFakeModuleFixture(depRoot, { cardKeyPrefix: 'reimpdep' });
+    const hostRoot = join(moduleTestDir, 'fake-host-reimport');
+    makeFakeModuleFixture(hostRoot, {
+      cardKeyPrefix: 'reimphost',
+      modules: [{ name: 'reimpdep', location: `file:${pathResolve(depRoot)}` }],
+    });
+
+    const projectDir = join(moduleTestDir, 'proj-reimport-orphan');
+    const handler = new Commands();
+    const create = await handler.command(
+      Cmd.create,
+      ['project', 'reimport-orphan-proj', 'riop'],
+      { projectPath: projectDir },
+    );
+    expect(create.statusCode).toBe(200);
+
+    const commands = new CommandManager(projectDir, {});
+    await commands.initialize();
+
+    // First import: host is declared, dep is installed transitively on disk.
+    await commands.modulesCmd.install(hostRoot);
+
+    const installedHost = join(projectDir, '.cards', 'modules', 'reimphost');
+    const installedDep = join(projectDir, '.cards', 'modules', 'reimpdep');
+    expect(existsSync(installedHost)).toBe(true);
+    expect(existsSync(installedDep)).toBe(true);
+    // Only top-level declarations appear in configuration.modules.
+    expect(
+      commands.project.configuration.modules.some(
+        (m) => m.name === 'reimphost',
+      ),
+    ).toBe(true);
+
+    // Upstream drops its transitive dep.
+    rewriteFakeModuleFixture(hostRoot, {
+      cardKeyPrefix: 'reimphost',
+      modules: [],
+    });
+
+    // Re-import via install (not updateAll).
+    await commands.modulesCmd.install(hostRoot);
+
+    // cleanOrphans must have run: dep folder gone, host still present.
+    expect(existsSync(installedHost)).toBe(true);
+    expect(existsSync(installedDep)).toBe(false);
+    // Host declaration is still intact.
+    expect(
+      commands.project.configuration.modules.some(
+        (m) => m.name === 'reimphost',
+      ),
+    ).toBe(true);
+  });
+
+  it('updateAll refreshes allModulePrefixes when a new transitive is pulled in', async () => {
+    // When `Modules.updateAll` pulls in a brand-new transitive (because the
+    // upstream started declaring a dep it previously did not), the project's
+    // cached `allModulePrefixes()` must immediately include that transitive's
+    // prefix without any manual refresh by the caller — the applier fires
+    // the refresh itself.
+    const depRoot = join(moduleTestDir, 'fake-new-dep');
+    makeFakeModuleFixture(depRoot, { cardKeyPrefix: 'newdep' });
+    const hostRoot = join(moduleTestDir, 'fake-new-host');
+    // Host initially declares no transitives.
+    makeFakeModuleFixture(hostRoot, {
+      cardKeyPrefix: 'newhost',
+      modules: [],
+    });
+
+    const projectDir = join(moduleTestDir, 'proj-new-transitive');
+    const commandHandler = new Commands();
+    const create = await commandHandler.command(
+      Cmd.create,
+      ['project', 'new-transitive-proj', 'ntrp'],
+      { projectPath: projectDir },
+    );
+    expect(create.statusCode).toBe(200);
+
+    const commands = new CommandManager(projectDir, {});
+    await commands.initialize();
+
+    await commands.modulesCmd.install(hostRoot);
+    // Initially only `newhost` is installed — no `newdep` yet.
+    expect(commands.project.allModulePrefixes()).toContain('newhost');
+    expect(commands.project.allModulePrefixes()).not.toContain('newdep');
+
+    // Upstream starts declaring the new transitive.
+    rewriteFakeModuleFixture(hostRoot, {
+      cardKeyPrefix: 'newhost',
+      modules: [{ name: 'newdep', location: `file:${pathResolve(depRoot)}` }],
+    });
+
+    await commands.modulesCmd.updateAll();
+
+    // The new transitive is installed on disk...
+    expect(existsSync(join(projectDir, '.cards', 'modules', 'newdep'))).toBe(
+      true,
+    );
+    // ...and immediately visible through the cached prefix list without any
+    // manual refresh call.
+    expect(commands.project.allModulePrefixes()).toContain('newdep');
+    expect(commands.project.allModulePrefixes()).toContain('newhost');
+  });
+
+  it('update with an override version that violates the declared range throws', async () => {
+    // Spec: the `update <name> <exact-version>` path must refuse a
+    // version that contradicts the declared range. Implemented via
+    // `validateVersionAgainstConstraints` in `Modules.update`.
+    const projectDir = join(moduleTestDir, 'proj-override-bad');
+    const commandHandler = new Commands();
+    const create = await commandHandler.command(
+      Cmd.create,
+      ['project', 'override-bad-proj', 'ovb'],
+      { projectPath: projectDir },
+    );
+    expect(create.statusCode).toBe(200);
+
+    const depRoot = join(moduleTestDir, 'fake-override-mod');
+    makeFakeModuleFixture(depRoot, { cardKeyPrefix: 'ovmod' });
+
+    const commands = new CommandManager(projectDir, {});
+    await commands.initialize();
+    await commands.modulesCmd.install(depRoot);
+
+    // Synthesise a declared range on the persisted module so the
+    // constraint-validation path fires. The caller-supplied override
+    // `2.0.0` violates `^1.0.0`.
+    const modSetting = commands.project.configuration.modules.find(
+      (m) => m.name === 'ovmod',
+    );
+    expect(modSetting).toBeDefined();
+    modSetting!.version = '^1.0.0';
+
+    await expect(
+      commands.modulesCmd.update('ovmod', undefined, '2.0.0'),
+    ).rejects.toThrow(/does not satisfy constraint '\^1\.0\.0'/);
+  });
+
+  it('update with an override version inside the declared range succeeds', async () => {
+    // Paired positive case for the override flow: `1.3.0` satisfies
+    // `^1.0.0`, so the constraint check passes and the install path
+    // completes end-to-end against a file source (which ignores the ref
+    // but still exercises the two-phase install).
+    const projectDir = join(moduleTestDir, 'proj-override-ok');
+    const commandHandler = new Commands();
+    const create = await commandHandler.command(
+      Cmd.create,
+      ['project', 'override-ok-proj', 'ovk'],
+      { projectPath: projectDir },
+    );
+    expect(create.statusCode).toBe(200);
+
+    const depRoot = join(moduleTestDir, 'fake-override-mod-ok');
+    makeFakeModuleFixture(depRoot, { cardKeyPrefix: 'ovkmod' });
+
+    const commands = new CommandManager(projectDir, {});
+    await commands.initialize();
+    await commands.modulesCmd.install(depRoot);
+
+    const modSetting = commands.project.configuration.modules.find(
+      (m) => m.name === 'ovkmod',
+    );
+    expect(modSetting).toBeDefined();
+    modSetting!.version = '^1.0.0';
+
+    // No throw: constraint check passes, file source fetch/install runs
+    // end-to-end. The persisted range stays at `^1.0.0` — the applier
+    // only writes back the declared range, never the resolved tag.
+    await commands.modulesCmd.update('ovkmod', undefined, '1.3.0');
+
+    const after = commands.project.configuration.modules.find(
+      (m) => m.name === 'ovkmod',
+    );
+    expect(after?.version).toBe(toVersionRange('^1.0.0'));
+    expect(existsSync(join(projectDir, '.cards', 'modules', 'ovkmod'))).toBe(
+      true,
+    );
+  });
+
+  it('update on a transitive-only module errors naming the parent', async () => {
+    // A transitive's version is owned by its parent's resolution — you
+    // cannot pin or update it directly. The error must surface the
+    // parent rather than misleadingly claim the module isn't part of
+    // the project.
+    const depRoot = join(moduleTestDir, 'fake-tup-dep');
+    makeFakeModuleFixture(depRoot, { cardKeyPrefix: 'tupdep' });
+    const hostRoot = join(moduleTestDir, 'fake-tup-host');
+    makeFakeModuleFixture(hostRoot, {
+      cardKeyPrefix: 'tuphost',
+      modules: [{ name: 'tupdep', location: `file:${pathResolve(depRoot)}` }],
+    });
+
+    const projectDir = join(moduleTestDir, 'proj-tup');
+    const commandHandler = new Commands();
+    const create = await commandHandler.command(
+      Cmd.create,
+      ['project', 'tup-proj', 'tupp'],
+      { projectPath: projectDir },
+    );
+    expect(create.statusCode).toBe(200);
+
+    const commands = new CommandManager(projectDir, {});
+    await commands.initialize();
+
+    await commands.modulesCmd.install(hostRoot);
+    const topLevelNames = commands.project.configuration.modules.map(
+      (m) => m.name,
+    );
+    expect(topLevelNames).toEqual(['tuphost']);
+    expect(existsSync(join(projectDir, '.cards', 'modules', 'tupdep'))).toBe(
+      true,
+    );
+
+    await expect(commands.modulesCmd.update('tupdep')).rejects.toThrow(
+      "Cannot update module 'tupdep' because it is required by 'tuphost'. Update the parent module(s) instead.",
+    );
+  });
+
+  it('update downgrade aborts before any disk change', async () => {
+    // A downgrade is unreachable by replay, so the resolver refuses it up
+    // front — before applyModules, so the installed tree stays untouched.
+    // File sources ignore the resolver's ref and report no remote versions,
+    // so the caller's exact-version override flows straight into the
+    // resolver while the installed version is seeded by hand.
+    const modRoot = join(moduleTestDir, 'fake-downgrade-mod');
+    makeFakeModuleFixture(modRoot, { cardKeyPrefix: 'dgmod' });
+
+    const projectDir = join(moduleTestDir, 'proj-downgrade');
+    const commandHandler = new Commands();
+    const create = await commandHandler.command(
+      Cmd.create,
+      ['project', 'downgrade-proj', 'dgp'],
+      { projectPath: projectDir },
+    );
+    expect(create.statusCode).toBe(200);
+
+    const commands = new CommandManager(projectDir, {});
+    await commands.initialize();
+    await commands.modulesCmd.install(modRoot);
+
+    // Seed the INSTALLED config with a version newer than the override
+    // target. The fixture's own config carries no version, so a
+    // successful applyModules would wipe this marker.
+    const installedConfigPath = join(
+      projectDir,
+      '.cards',
+      'modules',
+      'dgmod',
+      'cardsConfig.json',
+    );
+    const installedConfig = JSON.parse(
+      readFileSync(installedConfigPath, 'utf-8'),
+    );
+    installedConfig.version = '1.0.0';
+    writeFileSync(installedConfigPath, JSON.stringify(installedConfig));
+
+    await expect(
+      commands.modulesCmd.update('dgmod', undefined, '0.1.0'),
+    ).rejects.toThrow(/cannot downgrade from 1\.0\.0 to 0\.1\.0/i);
+
+    // The refusal fired during resolution: the seeded marker survives.
+    const afterConfig = JSON.parse(readFileSync(installedConfigPath, 'utf-8'));
+    expect(afterConfig.version).toBe('1.0.0');
+  });
+
+  it('updateAll writes the assumed 1.x into a version-less declaration', async () => {
+    // The shim's exit: a legacy declaration (written before version support
+    // existed) is converted the next time any module operation runs, even
+    // though this module has nowhere to move.
+    const projectDir = join(moduleTestDir, 'proj-backfill');
+    const handler = new Commands();
+    const create = await handler.command(
+      Cmd.create,
+      ['project', 'backfill-proj', 'bkf'],
+      { projectPath: projectDir },
+    );
+    expect(create.statusCode).toBe(200);
+
+    const depRoot = join(moduleTestDir, 'fake-backfill-mod');
+    makeFakeModuleFixture(depRoot, { cardKeyPrefix: 'bkfmod' });
+
+    const commands = new CommandManager(projectDir, {
+      autoSaveConfiguration: false,
+    });
+    await commands.initialize();
+    await commands.modulesCmd.install(depRoot);
+
+    // Reshape the persisted declaration into the legacy form: a versioned
+    // (git) source with no declared range. Importing from a file source
+    // already leaves the range off, so only the location has to change.
+    const declaration = commands.project.configuration.modules.find(
+      (m) => m.name === 'bkfmod',
+    );
+    expect(declaration).toBeDefined();
+    expect(declaration!.version).toBeUndefined();
+    declaration!.location = 'https://example.com/bkfmod.git';
+
+    // Seed the installed version so the module resolves to 1.0.0.
+    const installedConfigPath = join(
+      projectDir,
+      '.cards',
+      'modules',
+      'bkfmod',
+      'cardsConfig.json',
+    );
+    rewriteJson(installedConfigPath, (config) => {
+      config.version = '1.0.0';
+    });
+
+    mockEnsureModuleListUpToDate();
+    // 1.0.0 is the only tag, so the module stays exactly where it is.
+    vi.spyOn(GitManager, 'listRemoteVersionTags').mockResolvedValue(['1.0.0']);
+
+    await commands.modulesCmd.updateAll();
+
+    const persisted = JSON.parse(
+      readFileSync(
+        join(projectDir, '.cards', 'local', 'cardsConfig.json'),
+        'utf-8',
+      ),
+    );
+    const entry = persisted.modules.find(
+      (m: { name: string }) => m.name === 'bkfmod',
+    );
+    expect(entry.version).toBe('1.x');
+    // Still installed at the version it had — this is a declaration-only
+    // rewrite, not an update.
+    const installed = JSON.parse(readFileSync(installedConfigPath, 'utf-8'));
+    expect(installed.version).toBe('1.0.0');
+  });
+
+  // --- update replay end to end -------------------------------------
+  //
+  // File sources expose no remote versions (`listRemoteVersions` → []), and
+  // the resolver only pins `resolved.version` from an explicit override or
+  // from remote tags — never from the staged cardsConfig. So replay through
+  // `Modules.update` for a file source requires the caller's exact-version
+  // argument, exactly like the downgrade-conflict test above. The installed
+  // version is seeded by hand; the seal ships in the source's migrations/.
+
+  function rewriteJson(
+    path: string,
+    mutate: (value: Record<string, unknown>) => void,
+  ): void {
+    const value = JSON.parse(readFileSync(path, 'utf-8'));
+    mutate(value);
+    writeFileSync(path, JSON.stringify(value, null, 4));
+  }
+
+  /**
+   * Create a fresh host project, import the decision-records fixture (the
+   * private per-test copy under `moduleTestDir`), add a LOCAL card type
+   * referencing the module's `decision` workflow, and seed the installed
+   * module at version 1.0.0 (the fixture's own config carries no version,
+   * and its migrations folder is empty).
+   */
+  async function seedReplayHost(name: string, prefix: string) {
+    const moduleSource = join(moduleTestDir, 'valid', 'decision-records');
+    const projectDir = join(moduleTestDir, name);
+    const handler = new Commands();
+    const create = await handler.command(
+      Cmd.create,
+      ['project', name, prefix],
+      {
+        projectPath: projectDir,
+      },
+    );
+    expect(create.statusCode).toBe(200);
+
+    const commands = new CommandManager(projectDir, {});
+    await commands.initialize();
+    await commands.modulesCmd.install(moduleSource);
+
+    await commands.createCmd.createCardType(
+      `${prefix}/cardTypes/consumer`,
+      'decision/workflows/decision',
+    );
+
+    const installedConfigPath = join(
+      projectDir,
+      '.cards',
+      'modules',
+      'decision',
+      'cardsConfig.json',
+    );
+    rewriteJson(installedConfigPath, (config) => {
+      config.version = '1.0.0';
+    });
+
+    return { moduleSource, projectDir, commands, installedConfigPath };
+  }
+
+  /**
+   * Turn the SOURCE fixture into its 2.0.0 release: bump the config version
+   * and write one seal covering (0.0.0, 2.0.0] with the given released-format
+   * log lines. Callers apply the matching resource changes themselves so
+   * the source tree stays self-consistent (upstream already migrated).
+   */
+  function sealSourceRelease(moduleSource: string, lines: string[]): void {
+    const sourceLocal = join(moduleSource, '.cards', 'local');
+    rewriteJson(join(sourceLocal, 'cardsConfig.json'), (config) => {
+      config.version = '2.0.0';
+    });
+    writeSeals(join(sourceLocal, 'migrations'), [
+      { from: '0.0.0', to: '2.0.0', lines },
+    ]);
+  }
+
+  it('update replays a sealed workflow rename into the consumer (happy path)', async () => {
+    const { moduleSource, projectDir, commands, installedConfigPath } =
+      await seedReplayHost('replay-happy-proj', 'rhap');
+
+    // Upstream 2.0.0: workflow renamed decision -> resolution. The source
+    // tree carries the rename already applied (file moved, internal card
+    // type reference updated) plus the seal that records it.
+    const sourceLocal = join(moduleSource, '.cards', 'local');
+    const workflow = JSON.parse(
+      readFileSync(join(sourceLocal, 'workflows', 'decision.json'), 'utf-8'),
+    );
+    workflow.name = 'decision/workflows/resolution';
+    writeFileSync(
+      join(sourceLocal, 'workflows', 'resolution.json'),
+      JSON.stringify(workflow, null, 4),
+    );
+    rmSync(join(sourceLocal, 'workflows', 'decision.json'));
+    rewriteJson(join(sourceLocal, 'cardTypes', 'decision.json'), (cardType) => {
+      cardType.workflow = 'decision/workflows/resolution';
+    });
+    sealSourceRelease(moduleSource, [
+      logLine('resource_rename', 'decision/workflows/decision', {
+        type: 'workflows',
+        operation: {
+          name: 'change',
+          target: 'decision/workflows/decision',
+          to: 'decision/workflows/resolution',
+        },
+      }),
+    ]);
+
+    // plan -> applyModules -> replay -> final validation; must not throw.
+    await commands.modulesCmd.update('decision', undefined, '2.0.0');
+
+    // Module files landed: new version, renamed workflow file.
+    const afterConfig = JSON.parse(readFileSync(installedConfigPath, 'utf-8'));
+    expect(afterConfig.version).toBe('2.0.0');
+    const installedWorkflows = join(
+      projectDir,
+      '.cards',
+      'modules',
+      'decision',
+      'workflows',
+    );
+    expect(existsSync(join(installedWorkflows, 'resolution.json'))).toBe(true);
+    expect(existsSync(join(installedWorkflows, 'decision.json'))).toBe(false);
+
+    // The replay cascade rewrote the LOCAL consumer's reference on disk.
+    const consumer = JSON.parse(
+      readFileSync(
+        join(projectDir, '.cards', 'local', 'cardTypes', 'consumer.json'),
+        'utf-8',
+      ),
+    );
+    expect(consumer.workflow).toBe('decision/workflows/resolution');
+  }, 30000);
+
+  it('update replays a sealed workflow delete: dependent local card types and their cards are removed', async () => {
+    const { moduleSource, projectDir, commands, installedConfigPath } =
+      await seedReplayHost('replay-delete-proj', 'rdel');
+
+    // A template card and a project card of the consumer type; the delete
+    // cascade must remove the card type and both cards.
+    await commands.createCmd.createTemplate('rdel/templates/seed', '');
+    await commands.createCmd.addCards(
+      'rdel/cardTypes/consumer',
+      'rdel/templates/seed',
+    );
+    const created = await commands.createCmd.createCard('rdel/templates/seed');
+    expect(created.length).toBeGreaterThan(0);
+
+    // Upstream 2.0.0 deletes the workflow; its own card type moves to the
+    // remaining 'simple' workflow, keeping the module self-consistent.
+    const sourceLocal = join(moduleSource, '.cards', 'local');
+    rmSync(join(sourceLocal, 'workflows', 'decision.json'));
+    rewriteJson(join(sourceLocal, 'cardTypes', 'decision.json'), (cardType) => {
+      cardType.workflow = 'decision/workflows/simple';
+    });
+    sealSourceRelease(moduleSource, [
+      logLine('resource_delete', 'decision/workflows/decision', {
+        type: 'workflows',
+      }),
+    ]);
+
+    // A card type requires a workflow, so the replayed delete repairs the
+    // consumer by removal: card type and all of its cards. The project is
+    // then valid and the update succeeds.
+    await commands.modulesCmd.update('decision', undefined, '2.0.0');
+
+    const afterConfig = JSON.parse(readFileSync(installedConfigPath, 'utf-8'));
+    expect(afterConfig.version).toBe('2.0.0');
+    expect(
+      existsSync(
+        join(projectDir, '.cards', 'local', 'cardTypes', 'consumer.json'),
+      ),
+    ).toBe(false);
+    for (const card of created) {
+      expect(existsSync(join(projectDir, 'cardRoot', card.key))).toBe(false);
+    }
+    expect(
+      existsSync(
+        join(
+          projectDir,
+          '.cards',
+          'modules',
+          'decision',
+          'workflows',
+          'decision.json',
+        ),
+      ),
+    ).toBe(false);
+  }, 30000);
+
+  it('update surfaces ModuleValidationFailedError when the update leaves a dangling workflow reference', async () => {
+    const { moduleSource, projectDir, commands, installedConfigPath } =
+      await seedReplayHost('replay-invalid-proj', 'rinv');
+
+    // A project card of the MODULE's own card type. The delete cascade only
+    // repairs LOCAL card types, so this card's type survives pointing at the
+    // deleted workflow and card-level validation must fail.
+    const created = await commands.createCmd.createCard(
+      'decision/templates/decision',
+    );
+    expect(created.length).toBeGreaterThan(0);
+
+    // Upstream ships a broken 2.0.0: the workflow is deleted but the
+    // module's own card type still references it.
+    const sourceLocal = join(moduleSource, '.cards', 'local');
+    rmSync(join(sourceLocal, 'workflows', 'decision.json'));
+    sealSourceRelease(moduleSource, [
+      logLine('resource_delete', 'decision/workflows/decision', {
+        type: 'workflows',
+      }),
+    ]);
+
+    const error = await commands.modulesCmd
+      .update('decision', undefined, '2.0.0')
+      .catch((e) => e);
+
+    expect(error).toBeInstanceOf(ModuleValidationFailedError);
+    expect(error.steps).toEqual([
+      { modulePrefix: 'decision', fromVersion: '1.0.0', toVersion: '2.0.0' },
+    ]);
+    expect(error.message).toContain('decision/workflows/decision');
+
+    // Best-effort semantics: the module files DID update before validation
+    // failed — new version on disk, deleted workflow gone.
+    const afterConfig = JSON.parse(readFileSync(installedConfigPath, 'utf-8'));
+    expect(afterConfig.version).toBe('2.0.0');
+    expect(
+      existsSync(
+        join(
+          projectDir,
+          '.cards',
+          'modules',
+          'decision',
+          'workflows',
+          'decision.json',
+        ),
+      ),
+    ).toBe(false);
+  }, 30000);
+});
+
+describe('module schema version enforcement on import', () => {
+  const schemaTestDir = join(baseDir, 'tmp-import-schema-tests');
+  const hostPath = join(schemaTestDir, 'valid/minimal');
+  const modulePath = join(schemaTestDir, 'valid/decision-records');
+  const moduleConfigPath = join(
+    modulePath,
+    '.cards',
+    'local',
+    'cardsConfig.json',
+  );
+  const installedConfigPath = join(
+    hostPath,
+    '.cards',
+    'modules',
+    'decision',
+    'cardsConfig.json',
+  );
+
+  function setModuleSchemaVersion(version: number | undefined) {
+    const config = readJsonFileSync(moduleConfigPath) as Record<
+      string,
+      unknown
+    >;
+    if (version === undefined) {
+      delete config.schemaVersion;
+    } else {
+      config.schemaVersion = version;
+    }
+    writeFileSync(moduleConfigPath, JSON.stringify(config, null, 4));
+  }
+
+  beforeEach(async () => {
+    rmSync(schemaTestDir, { recursive: true, force: true });
+    mkdirSync(schemaTestDir, { recursive: true });
+    await copyDir('test/test-data', schemaTestDir);
+    mockEnsureModuleListUpToDate();
+  });
+
+  afterEach(() => {
+    vi.restoreAllMocks();
+    rmSync(schemaTestDir, { recursive: true, force: true });
+  });
+
+  it('migrates an older module before installing it', async () => {
+    setModuleSchemaVersion(4);
+    const commands = new CommandManager(hostPath);
+    await commands.initialize();
+    await commands.modulesCmd.install(modulePath);
+    const installed = readJsonFileSync(installedConfigPath) as {
+      schemaVersion: number;
+    };
+    expect(installed.schemaVersion).toBe(SCHEMA_VERSION);
+    commands.project.dispose();
+  });
+
+  it('refuses a module newer than the tool, before any disk change', async () => {
+    setModuleSchemaVersion(SCHEMA_VERSION + 1);
+    const commands = new CommandManager(hostPath);
+    await commands.initialize();
+    await expect(commands.modulesCmd.install(modulePath)).rejects.toThrow(
+      'Upgrade cyberismo',
+    );
+    expect(existsSync(installedConfigPath)).toBe(false);
+    commands.project.dispose();
+  });
+
+  it('refuses a module without a schema version', async () => {
+    setModuleSchemaVersion(undefined);
+    const commands = new CommandManager(hostPath);
+    await commands.initialize();
+    await expect(commands.modulesCmd.install(modulePath)).rejects.toThrow(
+      "no 'schemaVersion'",
+    );
+    commands.project.dispose();
+  });
+});
